app/config.py
import os
from dotenv import load_dotenv


def load_config(app):
    load_dotenv()
    app.config["OPENAI_API_KEY"] = os.getenv("OPENAI_API_KEY", "")
    app.config["HOST"] = os.getenv("HOST", "0.0.0.0")
    app.config["PORT"] = int(os.getenv("PORT", 5000))
