requirements.txt
Flask>=3.0
python-dotenv>=1.0
Pillow>=10.4
openai>=1.40
pydantic>=2.8
pytesseract>=0.3.10
PyMuPDF>=1.24.8          # PDF rasterization (no external poppler needed)
opencv-python-headless>=4.8
numpy>=1.26
cryptography>=42.0.0
gunicorn>=22.0